rename: 토스트 컴포넌트 이름 수정
