+++ src/hooks/useToast.ts
@@ -1,7 +1,7 @@
 // Inspired by react-hot-toast library
 import * as React from "react"
 
-import type { ToastActionElement, ToastProps } from "@/components/ui/Toast"
+import type { ToastActionElement, ToastProps } from "@/components/ui/ToastUI"
 
 const TOAST_LIMIT = 1
 const TOAST_REMOVE_DELAY = 1000000
